+++ config/events-web.php
@@ -2,4 +2,11 @@
 
 declare(strict_types=1);
 
-return [];
+use App\Timer;
+use Yiisoft\Yii\Http\Event\ApplicationStartup;
+
+return [
+    ApplicationStartup::class => [
+        static fn(Timer $timer) => $timer->start('overall'),
+    ],
+];